requirements.txt
aiosignal==1.3.1
alabaster==0.7.12
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==3.0.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
botorch==0.8.1
brotlipy==0.7.0
build==1.0.3
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==2.0.4
ciso8601==2.3.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cryptography==41.0.3
cycler==0.11.0
datasets==2.8.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dgl==1.1.2.post1
dglgo==0.0.2
dill==0.3.6
distlib==0.3.7
docutils==0.18.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.4
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
gpytorch==1.9.1
huggingface-hub==0.11.1
identify==2.5.30
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
ipykernel==6.19.4
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter-contrib-core==0.4.2
jupyter_core==5.1.2
jupyter-events==0.5.0
jupyter-nbextensions-configurator==0.6.1
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
linear-operator==0.3.0
littleutils==0.2.2
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
miniful==0.0.6
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.4
multipledispatch==0.6.0
multiprocess==0.70.14
mypy==1.5.1
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.8.8
networkx-stubs==0.0.1
nodeenv==1.8.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
numpydoc==1.5.0
ogb==1.3.5
opt-einsum==3.3.0
outdated==0.2.2
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.2.1
platformdirs==3.11.0
Platypus-Opt==1.1.0
ply==3.11
pre-commit==3.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pyDOE==0.3.8
pyFUME==0.2.25
Pygments==2.14.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyro-api==0.1.2
pyro-ppl==1.8.4
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch-warmup==0.1.1
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
rdkit-pypi==2022.9.3
regex==2022.10.31
requests==2.28.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==68.0.0
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
sympy==1.12
tensorboardX==2.5.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typer==0.7.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
virtualenv==20.24.5
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
wheel==0.41.2
widgetsnbextension==4.0.5
xxhash==3.2.0
yarl==1.8.2
zipp==3.11.0
